Fixing conflixts with migration
