--- backend/framework/alembic/versions/31a13e8e1ed2_add_monster_majocr_table.py
@@ -1,37 +1,0 @@
-"""Add monster_majocr table
-
-Revision ID: 31a13e8e1ed2
-Revises: aa75144a150e
-Create Date: 2025-08-22 11:24:58.311445
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '31a13e8e1ed2'
-down_revision: Union[str, Sequence[str], None] = 'aa75144a150e'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('monster_majocr',
-    sa.Column('index', sa.String(), nullable=False),
-    sa.Column('name', sa.String(), nullable=False),
-    sa.Column('data', sa.JSON(), nullable=False),
-    sa.PrimaryKeyConstraint('index')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('monster_majocr')
-    # ### end Alembic commands ###
