--- backend/framework/alembic/versions/5838f470f73b_castroulloaaaron_generate_dnd_caching_.py
@@ -1,42 +1,0 @@
-"""castroulloaaaron generate dnd caching tables
-
-Revision ID: 5838f470f73b
-Revises: aa75144a150e
-Create Date: 2025-08-23 17:44:25.420194
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '5838f470f73b'
-down_revision: Union[str, Sequence[str], None] = 'aa75144a150e'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('castroulloaaaron_allmonsters',
-                    sa.Column('id', sa.Integer(), nullable=False),
-                    sa.Column('json_data', sa.JSON(), nullable=False),
-                    sa.PrimaryKeyConstraint('id')
-                    )
-    op.create_table('castroulloaaaron_monsters',
-                    sa.Column('id', sa.String(), nullable=False),
-                    sa.Column('json_data', sa.JSON(), nullable=False),
-                    sa.PrimaryKeyConstraint('id')
-                    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('castroulloaaaron_monsters')
-    op.drop_table('castroulloaaaron_allmonsters')
-    # ### end Alembic commands ###
