--- backend/framework/alembic/versions/a5c414a8924a_damurill_generate_monsters.py
@@ -1,42 +1,0 @@
-"""damurill_generate_monsters
-
-Revision ID: a5c414a8924a
-Revises: 5838f470f73b
-Create Date: 2025-08-27 10:32:54.376510
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'a5c414a8924a'
-down_revision: Union[str, Sequence[str], None] = '5838f470f73b'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('dmuri_monsters',
-    sa.Column('index', sa.String(length=100), nullable=False),
-    sa.Column('data', sa.JSON(), nullable=False),
-    sa.PrimaryKeyConstraint('index')
-    )
-    op.create_table('dmuri_monsters_list',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('data', sa.JSON(), nullable=False),
-    sa.PrimaryKeyConstraint('id')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('dmuri_monsters_list')
-    op.drop_table('dmuri_monsters')
-    # ### end Alembic commands ###
