+++ backend/framework/alembic/versions/a6e5d0a36ec3_damurill_generate_monsters.py
@@ -1,0 +1,74 @@
+"""damurill_generate_monsters
+
+Revision ID: a6e5d0a36ec3
+Revises: 
+Create Date: 2025-08-27 10:47:22.470433
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = 'a6e5d0a36ec3'
+down_revision: Union[str, Sequence[str], None] = None
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    """Upgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('castroulloaaaron_allmonsters',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('json_data', sa.JSON(), nullable=False),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_table('castroulloaaaron_monsters',
+    sa.Column('id', sa.String(), nullable=False),
+    sa.Column('json_data', sa.JSON(), nullable=False),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_table('dmuri_monsters',
+    sa.Column('index', sa.String(length=100), nullable=False),
+    sa.Column('data', sa.JSON(), nullable=False),
+    sa.PrimaryKeyConstraint('index')
+    )
+    op.create_table('dmuri_monsters_list',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('data', sa.JSON(), nullable=False),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_table('monster_list_majocr',
+    sa.Column('index', sa.String(), nullable=False),
+    sa.Column('name', sa.String(), nullable=False),
+    sa.Column('url', sa.String(), nullable=False),
+    sa.PrimaryKeyConstraint('index')
+    )
+    op.create_table('monster_majocr',
+    sa.Column('index', sa.String(), nullable=False),
+    sa.Column('name', sa.String(), nullable=False),
+    sa.Column('data', sa.JSON(), nullable=False),
+    sa.PrimaryKeyConstraint('index')
+    )
+    op.create_table('motivational_phrases',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('phrase', sa.String(), nullable=True),
+    sa.PrimaryKeyConstraint('id')
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    """Downgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_table('motivational_phrases')
+    op.drop_table('monster_majocr')
+    op.drop_table('monster_list_majocr')
+    op.drop_table('dmuri_monsters_list')
+    op.drop_table('dmuri_monsters')
+    op.drop_table('castroulloaaaron_monsters')
+    op.drop_table('castroulloaaaron_allmonsters')
+    # ### end Alembic commands ###
