--- backend/framework/alembic/versions/aa75144a150e_adding_motivation.py
@@ -1,36 +1,0 @@
-"""Adding motivation
-
-Revision ID: aa75144a150e
-Revises: 
-Create Date: 2025-07-20 16:12:34.938097
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'aa75144a150e'
-down_revision: Union[str, Sequence[str], None] = None
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('motivational_phrases',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('phrase', sa.String(), nullable=True),
-    sa.PrimaryKeyConstraint('id')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('motivational_phrases')
-    # ### end Alembic commands ###
