--- backend/framework/alembic/versions/cd4474e7e340_add_monster_list_majocr_table.py
@@ -1,37 +1,0 @@
-"""Add monster_list_majocr table
-
-Revision ID: cd4474e7e340
-Revises: 31a13e8e1ed2
-Create Date: 2025-08-25 18:23:03.419033
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'cd4474e7e340'
-down_revision: Union[str, Sequence[str], None] = '31a13e8e1ed2'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('monster_list_majocr',
-    sa.Column('index', sa.String(), nullable=False),
-    sa.Column('name', sa.String(), nullable=False),
-    sa.Column('url', sa.String(), nullable=False),
-    sa.PrimaryKeyConstraint('index')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('monster_list_majocr')
-    # ### end Alembic commands ###
